backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML Libraries
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Vector Database and Embeddings
weaviate-client==3.25.3
pinecone-client==2.2.4
sentence-transformers==2.2.2

# MCP (Model Context Protocol) - Core Integration
# mcp==0.1.0  # Not available yet, using custom implementation
httpx==0.25.2
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==3.4.8
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
